gol 8-bit states pre-optimized
