+++ constants/data.constant.go
@@ -1,6 +1,7 @@
 package constants
 
-const GAME_TYPE_GOL = "Game of Life"
-const GAME_TYPE_LTL = "Larger than Life"
-
-const RANDOM = "Random" // GoL & LtL Preset
+const (
+	GAME_TYPE_GOL         = "Game of Life"
+	GAME_TYPE_LTL         = "Larger than Life"
+	GAME_TYPE_SMOOTH_LIFE = "Smooth Life"
+)
