+++ main.go
@@ -4,32 +4,35 @@ import (
 	"github.com/ebitenui/ebitenui"
 	"github.com/hajimehoshi/ebiten/v2"
 	"github.com/tasnimzotder/artificial-life/settings"
-	"github.com/tasnimzotder/artificial-life/ui2"
+	"github.com/tasnimzotder/artificial-life/ui"
 	"log"
 )
 
 func main() {
 	gameSettings := settings.NewGameSettings()
-	uiWidgets := ui2.NewUiWidget(gameSettings)
+	uiWidgets := ui.NewUiWidget(gameSettings)
 
 	eui := &ebitenui.UI{
 		Container: uiWidgets.RootContainer,
 	}
 
-	g := &ui2.Game{
+	g := &ui.Game{
 		Settings: gameSettings,
 		World: settings.NewWorld(
 			gameSettings.WorldWidth,
 			gameSettings.WorldHeight,
-			(gameSettings.WorldWidth*gameSettings.WorldHeight)/2,
+			gameSettings,
 		),
 		UI: eui,
 	}
 
-	ebiten.SetWindowSize(ui2.ScreenWidth, ui2.ScreenHeight)
+	ebiten.SetWindowSize(ui.ScreenWidth, ui.ScreenHeight)
 	ebiten.SetWindowTitle("Artificial Life")
 	ebiten.SetVsyncEnabled(true)
+	ebiten.SetTPS(100)
 	//ebiten.SetFullscreen(true)
+
+	g.UpdateActualSpeed()
 
 	if err := ebiten.RunGame(g); err != nil {
 		log.Fatalf("Ebiten run game error: %v", err)
