+++ settings/game.setting.go
@@ -1,0 +1,69 @@
+package settings
+
+import "github.com/tasnimzotder/artificial-life/data"
+
+type GameSettings struct {
+	T         uint
+	GameType  string
+	GameTypes []string
+
+	Presets []data.PresetType
+	Preset  string
+
+	WrapAround  bool
+	IsPaused    bool
+	IsRunning   bool
+	IsReset     bool
+	Replay      bool
+	DesiredTPS  int
+	ZoomLevel   int
+	WorldWidth  int
+	WorldHeight int
+
+	//	ltl
+	LtLRule string
+	Radius  uint8
+}
+
+func NewGameSettings() *GameSettings {
+	gs := &GameSettings{}
+	setDefaultSettings(gs)
+
+	gs.IsReset = true
+	gs.Preset = "Random"
+	gs.GameType = gs.GameTypes[0]
+
+	gs.ZoomLevel = 10
+	gs.DesiredTPS = 2
+	gs.Radius = 1
+	gs.LtLRule = "Life"
+
+	return gs
+}
+
+func (gs *GameSettings) ResetSettings() {
+	setDefaultSettings(gs)
+	gs.IsReset = false
+	gs.Replay = false
+}
+
+func setDefaultSettings(gs *GameSettings) {
+	gs.T = 0
+	gs.WrapAround = true
+	gs.IsPaused = true
+	gs.IsReset = false
+
+	//gs.DesiredTPS = 10
+	//gs.ZoomLevel = 15
+	//gs.WorldWidth = 60 * 2
+	//gs.WorldHeight = 70 * 2
+	gs.WorldWidth = 512 * 1
+	gs.WorldHeight = 512 * 1
+
+	gs.GameTypes = []string{
+		"Game of Life",
+		//"Larger than Life",
+		//"Smooth Life",
+	}
+	gs.Presets = data.GetAllGoLPresets()
+}
