+++ settings/world-smoothlife.setting.go
@@ -1,0 +1,20 @@
+package settings
+
+const (
+	Ra     = 21
+	AlphaN = 0.028
+	AlphaM = 0.147
+	B1     = 0.278
+	B2     = 0.365
+	D1     = 0.267
+	D2     = 0.445
+	DT     = 0.005
+)
+
+func (w *World) InitRandomSmoothLife(gameType string, maxInitialLiveCells int) {
+	// todo: implement this
+}
+
+func (w *World) NextGenerationSmoothLife(gs *GameSettings) {
+	// todo: implement this
+}
